create AddAppAuthintication to contain all authintication code in program cs to make code clean
