+++ Mango.Services.CouponAPI/Extensions/WebApplicationBuilderExtensions.cs
@@ -1,0 +1,41 @@
+﻿using Microsoft.AspNetCore.Authentication.JwtBearer;
+using Microsoft.IdentityModel.Tokens;
+using System.Runtime.CompilerServices;
+using System.Text;
+
+namespace Mango.Services.CouponAPI.Extensions
+{
+    public static class WebApplicationBuilderExtensions
+    {
+        public static WebApplicationBuilder AddAppAuthintication(this WebApplicationBuilder builder)
+        {
+
+            var secret = builder.Configuration.GetValue<string>("ApiSetting:SecretKey");
+            var issuer = builder.Configuration.GetValue<string>("ApiSetting:Issuer");
+            var audience = builder.Configuration.GetValue<string>("ApiSetting:Audience");
+
+
+            var key = Encoding.ASCII.GetBytes(secret);
+
+
+            builder.Services.AddAuthentication(x => {
+                x.DefaultAuthenticateScheme = JwtBearerDefaults.AuthenticationScheme;
+                x.DefaultChallengeScheme = JwtBearerDefaults.AuthenticationScheme;
+
+            }).AddJwtBearer(x => {
+                x.RequireHttpsMetadata = false;
+                x.SaveToken = true;
+                x.TokenValidationParameters = new TokenValidationParameters
+                {
+                    ValidateIssuerSigningKey = true,
+                    IssuerSigningKey = new SymmetricSecurityKey(key),
+                    ValidateIssuer = true,
+                    ValidIssuer = issuer,
+                    ValidateAudience = true,
+                    ValidAudience = audience
+                };
+            });
+            return builder;
+        }
+    }
+}
