+++ Mango.Services.CouponAPI/Program.cs
@@ -1,5 +1,6 @@
 using Mango.Services.CouponAPI;
 using Mango.Services.CouponAPI.Data;
+using Mango.Services.CouponAPI.Extensions;
 using Mango.Services.CouponAPI.Repository;
 using Mango.Services.CouponAPI.Repository.IRepository;
 using Microsoft.AspNetCore.Authentication.JwtBearer;
@@ -49,31 +50,8 @@ builder.Services.AddAutoMapper(typeof(MappingConfig));
 builder.Services.AddScoped<ICouponRepository,CouponRepository>();
 builder.Services.AddScoped<IUnitOfWork,UnitOfWork>();
 
-var secret = builder.Configuration.GetValue<string>("ApiSetting:SecretKey");
-var issuer = builder.Configuration.GetValue<string>("ApiSetting:Issuer");
-var audience = builder.Configuration.GetValue<string>("ApiSetting:Audience");
 
-
-var key = Encoding.ASCII.GetBytes(secret);
-
-
-builder.Services.AddAuthentication(x => {
-    x.DefaultAuthenticateScheme = JwtBearerDefaults.AuthenticationScheme;
-    x.DefaultChallengeScheme = JwtBearerDefaults.AuthenticationScheme;
-
-}).AddJwtBearer(x => {
-    x.RequireHttpsMetadata = false;
-    x.SaveToken = true;
-    x.TokenValidationParameters = new TokenValidationParameters
-    {
-        ValidateIssuerSigningKey = true,
-        IssuerSigningKey = new SymmetricSecurityKey(key),
-        ValidateIssuer = true,
-        ValidIssuer = issuer,
-        ValidateAudience = true,
-        ValidAudience = audience
-    };
-});
+builder.AddAppAuthintication(); // extension method that i created to make code clean
 
 builder.Services.AddAuthorization();
